Version candidate for most recent compile date
- Extended the ClassTimestampAnalyzer to also provide a version candidate representing the most recent compile date (i.e. creation or modification date) of all classes contained in the jar.
I came across some jars, which had - compared to earlier versions of the same jars - only part of the classes replaced with different class file versions with a more recent compile date. In this case it is handy to have a version suggestion for this more recent compile date in the tool, too. Otherwise you would have to open the jar in a zip tool and find the most recent compile date manually.